LayoutTests/http/tests/security/contentSecurityPolicy/media-src-track-block-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/track.vtt because it does not appear in the media-src directive of the Content Security Policy.
ALERT: PASS
